+++ zeta_bot/core.py
@@ -3,7 +3,9 @@ import sys
 import os
 import asyncio
 import requests
+import platform
 from typing import Any, Union
+from discord.ui import Button, View
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from apscheduler.triggers.cron import CronTrigger
 
@@ -13,6 +15,7 @@ from zeta_bot import (
     utils,
     setting,
     log,
+    playlist,
     member,
     guild,
     help,
@@ -27,6 +30,12 @@ author = "炤铭Zeta (31Zeta)"
 python_path = sys.executable
 pycord_version = discord.__version__
 update_time = "2023.**.**"
+
+# 系统检测
+if platform.system().lower() == "windows":
+    ffmpeg_path = "./bin/ffmpeg.exe"
+else:
+    ffmpeg_path = "./bin/ffmpeg"
 
 # 多语言模块
 lang = language.Lang()
@@ -52,6 +61,8 @@ error_log_path = f"./logs/{log_name_time}_errors.log"
 log_path = f"./logs/{log_name_time}.log"
 logger = log.Log(error_log_path, log_path, setting.value("log"))
 
+logger.rp("程序启动", "[系统]")
+
 # 设置用户和Discord服务器管理
 utils.create_folder("./data")
 member_lib = member.MemberLibrary()
@@ -65,8 +76,6 @@ audio_lib_main = file_management.AudioFileLibrary(
     "主音频文件库",
     setting.value("audio_library_storage_size")
 )
-
-logger.rp("程序启动", "[系统]")
 
 
 def start(mode: str) -> None:
@@ -223,7 +232,18 @@ async def auto_reboot_reminder():
         if voice_client is not None:
             await current_guild.text_channels[0].send(f"注意：将在{ar_time}时自动重启")
 
-import time
+
+async def send_edit(ctx: discord.ApplicationContext, response, content: str, view=None) -> None:
+    """
+    如果<response>可以被编辑（类型为discord.InteractionMessage），则将<response>编辑为<content>
+    否则使用<ctx>发送<content>
+    """
+    if isinstance(response, discord.InteractionMessage):
+        await response.edit(content, view=view)
+    else:
+        await ctx.send(content, view=view)
+
+
 @bot.command(description="[管理员] 测试指令")
 async def debug(ctx: discord.ApplicationContext):
     """
@@ -247,20 +267,22 @@ async def debug(ctx: discord.ApplicationContext):
     # end_test = await ctx.send("测试结果已打印")
     # print(type(end_test))
 
-    start_test = await ctx.respond("开始测试")
-    print(type(start_test))
+    await ctx.respond("开始测试")
+    # print(type(start_test))
+    #
+    # if isinstance(start_test, discord.Interaction):
+    #     print("转换类型")
+    #     start_test = await start_test.original_response()
+    # print(type(start_test))
+    #
+    # time.sleep(3)
+    #
+    # if isinstance(start_test, discord.InteractionMessage):
+    #     print("类型符合，开始修改")
+    #     new_msg = await start_test.edit(content="测试回复已被修改")
+    #     print(type(new_msg))
 
-    if isinstance(start_test, discord.Interaction):
-        print("转换类型")
-        start_test = await start_test.original_response()
-    print(type(start_test))
-
-    time.sleep(3)
-
-    if isinstance(start_test, discord.InteractionMessage):
-        print("类型符合，开始修改")
-        new_msg = await start_test.edit(content="测试回复已被修改")
-        print(type(new_msg))
+    await ctx.respond("二次发送")
 
 
 @bot.command(description="关于Zeta-Discord机器人")
@@ -313,7 +335,7 @@ async def broadcast(ctx: discord.ApplicationContext, message) -> None:
 
 
 @bot.command(description=f"让 {bot_name} 加入语音频道")
-async def join(ctx: discord.ApplicationContext, channel_name=None) -> bool:
+async def join(ctx: discord.ApplicationContext, channel_name=None, function_call: bool = False) -> bool:
     """
     让机器人加入指令发送者所在的语音频道并发送提示\n
     如果机器人已经加入一个频道则转移到新频道并发送提示
@@ -321,9 +343,10 @@ async def join(ctx: discord.ApplicationContext, channel_name=None) -> bool:
 
     :param ctx: 指令原句
     :param channel_name: 要加入的频道名称
+    :param function_call 该指令是否是由其他函数调用
     :return: 布尔值，是否成功加入频道
     """
-    if not await command_check(ctx):
+    if not function_call and not await command_check(ctx):
         return False
 
     guild_lib.check(ctx)
@@ -359,13 +382,13 @@ async def join(ctx: discord.ApplicationContext, channel_name=None) -> bool:
     # 机器人未在任何语音频道的情况
     if voice_client is None:
         await channel.connect()
-        await ctx.respond(f"加入语音频道：-> ***{channel.name}***")
+        await ctx.respond(f"加入语音频道：->  ***{channel.name}***")
 
     # 机器人已经在一个频道的情况
     else:
         previous_channel = voice_client.channel
         await voice_client.move_to(channel)
-        await ctx.respond(f"转移语音频道：***{previous_channel}*** -> ***{channel.name}***")
+        await ctx.respond(f"转移语音频道：***{previous_channel}***  ->  ***{channel.name}***")
 
     logger.rp(f"加入语音频道：{channel.name}", ctx.guild)
     return True
@@ -427,12 +450,12 @@ async def play(ctx: discord.ApplicationContext, link=None) -> None:
     # 检测机器人是否已经加入语音频道
     if ctx.guild.voice_client is None:
         logger.rp("机器人未在任何语音频道中，尝试加入语音频道", ctx.guild)
-        join_result = await join(ctx)
+        join_result = await join(ctx, function_call=True)
         if not join_result:
             return
 
     # 尝试恢复之前被停止的播放
-    # await resume(ctx, play_call=True)
+    await resume(ctx, play_call=True)
 
     if link is None:
         logger.rp("用户未输入任何参数，指令无效", ctx.guild)
@@ -441,107 +464,49 @@ async def play(ctx: discord.ApplicationContext, link=None) -> None:
 
     # 检查输入的URL属于哪个网站
     source = utils.check_url_source(link)
-    logger.rp(f"检测输入的链接为类型：{source}", ctx.guild)
 
     # 如果指令中包含链接则提取链接
     if source is not None:
         link = utils.get_url_from_str(link, source)
 
     # URL属于Bilibili
-    if source == "bili_bvid" or source == "bili_url" or source == "bili_short_url":
-        # 如果是Bilibili短链则获取重定向链接
-        if source == "bili_short_url":
-            try:
-                link = utils.get_redirect_url(link)
-            except requests.exceptions.InvalidSchema:
-                await ctx.respond("链接异常")
-                logger.rp(f"链接重定向失败", ctx.guild)
+    if source == "bilibili_bvid" or source == "bilibili_url" or source == "bilibili_short_url":
+        loading_msg = await ctx.respond("正在加载Bilibili音频信息")
+        await play_bilibili(ctx, source, link, response=loading_msg)
 
-            logger.rp(f"获取的重定向链接为 {link}", ctx.guild)
-
-        # 如果是URl则转换成BV号
-        if source == "bili_url" or source == "bili_short_url":
-            bvid = utils.get_bvid_from_url(link)
-            if bvid is None:
-                logger.rp(f"{link} 为无效的链接", ctx.guild)
-                await ctx.respond("无效的Bilibili链接")
-                return
-        else:
-            bvid = link
-
-        # 获取Bilibili视频信息
-        info_dict = await bilibili.get_info(bvid)
-
-        # 单一视频 bili_single
-        if info_dict["videos"] == 1 and "ugc_season" not in info_dict:
-            loading_msg = await ctx.respond("正在加载Bilibili歌曲")
-            loaded_audio = await play_bili(ctx, info_dict, "bili_single", 0)
-
-            # TODO 测试用，非最终显示语句
-            await loading_msg.edit(f"测试：加载 {loaded_audio._title}")
-
-        # 合集视频 bili_collection
-        elif "ugc_season" in info_dict:
-            await play_bili(ctx, info_dict, "bili_single", 0)
-
-            collection_title = info_dict["ugc_season"]["title"]
-            message = f"此视频包含在合集 **{collection_title}** 中, 是否要查看此合集？\n"
-            view = CheckBiliCollectionView(ctx, info_dict)
-            await ctx.respond(message, view=view)
-
-        # 分P视频 bili_p
-        else:
-            message = "这是一个分p视频, 请选择要播放的分p:\n"
-            for item in info_dict["pages"]:
-                p_num = item["page"]
-                p_title = item["part"]
-                p_time_str = utils.convert_duration_to_time_str(item["duration"])
-                message = message + f"    **[{p_num}]** {p_title}  [{p_time_str}]\n"
-
-            menu_list = make_menu_list_10(message)
-            view = EpisodeSelectView(ctx, "bili_p", info_dict, menu_list)
-            await ctx.respond(f"{menu_list[0]}\n第[1]页，"
-                              f"共[{len(menu_list)}]页\n已输入：",
-                              view=view)
-
-    elif source == "ytb_url":
-
-        loading_msg = await ctx.send("正在获取Youtube视频信息")
-        url_type, info_dict = youtube.get_info(link)
-        await loading_msg.delete()
-
-        # 单一视频 ytb_single
-        if url_type == "ytb_single":
-            loading_msg = await ctx.respond("正在加载Youtube歌曲")
-            await play_ytb(ctx, link, info_dict, url_type)
-            # await loading_msg.delete()
-
-        # 播放列表 ytb_playlist
-        else:
-            message = "这是一个播放列表, 请选择要播放的集数:\n"
-            counter = 1
-            for item in info_dict["entries"]:
-                ep_num = counter
-                ep_title = item["fulltitle"]
-                ep_time_str = utils.convert_duration_to_time_str(item["duration"])
-                message = message + f"    **[{ep_num}]** {ep_title}  [{ep_time_str}]\n"
-                counter += 1
-
-            menu_list = make_menu_list_10(message)
-            view = EpisodeSelectView(ctx, "ytb_playlist", info_dict, menu_list)
-            await ctx.respond(f"{menu_list[0]}\n第[1]页，"
-                              f"共[{len(menu_list)}]页\n已输入：",
-                              view=view)
+    # elif source == "ytb_url":
+    #
+    #     loading_msg = await ctx.send("正在获取Youtube视频信息")
+    #     url_type, info_dict = youtube.get_info(link)
+    #     await loading_msg.delete()
+    #
+    #     # 单一视频 ytb_single
+    #     if url_type == "ytb_single":
+    #         loading_msg = await ctx.respond("正在加载Youtube歌曲")
+    #         await play_ytb(ctx, link, info_dict, url_type)
+    #         # await loading_msg.delete()
+    #
+    #     # 播放列表 ytb_playlist
+    #     else:
+    #         message = "这是一个播放列表, 请选择要播放的集数:\n"
+    #         ep_info_list = []
+    #         for item in info_dict["entries"]:
+    #             ep_title = item["fulltitle"]
+    #             ep_time_str = utils.convert_duration_to_time_str(item["duration"])
+    #             ep_info_list.append((ep_title, ep_time_str))
+    #
+    #         menu_list = utils.make_playlist_page(ep_info_list, 10, indent=4)
+    #         view = EpisodeSelectView(ctx, "ytb_playlist", info_dict, menu_list)
+    #         await ctx.respond(f"{menu_list[0]}\n第[1]页，共[{len(menu_list)}]页\n已输入：", view=view)
 
     else:
-        if link != "N/A":
+        if link is not None:
             await search_ytb(ctx, link)
 
 
-async def play_audio(
-        ctx: discord.ApplicationContext, target_audio: audio.Audio,
-        response: Union[discord.Interaction, discord.InteractionMessage, None] = None
-) -> None:
+async def play_audio(ctx: discord.ApplicationContext, target_audio: audio.Audio,
+                     response: Union[discord.Interaction, discord.InteractionMessage, None] = None,
+                     function_call: bool = False) -> None:
     """
     在<ctx>中的音频端播放音频<target_audio>，如果<single>为True则发送单曲加入成功通知
     <response>为用来编辑的加载信息，如果为None则发送新的通知
@@ -549,6 +514,7 @@ async def play_audio(
     :param ctx: 指令原句
     :param target_audio: 需要播放的音频
     :param response: 用于编辑的加载信息
+    :param function_call 该指令是否是由其他函数调用
     """
     # 将Interaction转换为InteractionMessage
     if isinstance(response, discord.Interaction):
@@ -560,22 +526,20 @@ async def play_audio(
     voice_client.play(
         discord.PCMVolumeTransformer(
             discord.FFmpegPCMAudio(
-                executable=setting.value("ffmpeg_path"), source=target_audio.get_path()
+                executable=ffmpeg_path, source=target_audio.get_path()
             )
         ),
         after=lambda e: asyncio.run_coroutine_threadsafe(play_next(ctx), bot.loop)
     )
     voice_client.source.volume = current_guild.get_voice_volume() / 100.0
 
-    logger.rp(
-        f"开始播放：{target_audio.get_title()} [{target_audio.get_time_str()}] 路径: {target_audio.get_path()}",
-        ctx.guild
-    )
+    if not function_call:
+        logger.rp(
+            f"开始播放：{target_audio.get_title()} [{target_audio.get_time_str()}] 路径: {target_audio.get_path()}",
+            ctx.guild
+        )
 
-    if isinstance(response, discord.InteractionMessage):
-        await response.edit(f"正在播放：**{target_audio.get_title()} [{target_audio.get_time_str()}]**")
-    else:
-        await ctx.send(f"正在播放：**{target_audio.get_title()} [{target_audio.get_time_str()}]**")
+        await send_edit(ctx, response, f"正在播放：**{target_audio.get_title()} [{target_audio.get_time_str()}]**")
 
 
 async def play_next(ctx: discord.ApplicationContext) -> None:
@@ -602,18 +566,76 @@ async def play_next(ctx: discord.ApplicationContext) -> None:
         await ctx.send("播放队列已结束")
 
 
-async def play_bili(ctx: discord.ApplicationContext, info_dict, download_type="bili_single", num_option=0,
-                    response: Union[discord.Interaction, discord.InteractionMessage, None] = None):
+async def play_bilibili(ctx: discord.ApplicationContext, source, link,
+                        response: Union[discord.Interaction, discord.InteractionMessage, None] = None):
     """
     下载并播放来自Bilibili的视频的音频
 
     :param ctx: 指令原句
-    :param info_dict: 目标的信息字典（使用bili_getinfo提取）
-    :param download_type: 下载模式（"bili_single"或"bili_p"）
-    :param num_option: 下载分集号（从0开始，默认为0即合集第1视频或者第1p）
+    :param source: 链接类型
+    :param link: 链接
     :param response: 用于编辑的加载信息
     :return: 播放的音频
     """
+    # 将Interaction转换为InteractionMessage
+    if isinstance(response, discord.Interaction):
+        response = await response.original_response()
+
+    # 如果是Bilibili短链则获取重定向链接
+    if source == "bilibili_short_url":
+        try:
+            link = utils.get_redirect_url_bilibili(link)
+        except requests.exceptions.InvalidSchema:
+            await send_edit(ctx, response, "链接异常")
+            logger.rp(f"链接重定向失败", ctx.guild)
+            return
+
+        logger.rp(f"获取的重定向链接为 {link}", ctx.guild)
+
+    # 如果是URl则转换成BV号
+    if source == "bilibili_url" or source == "bilibili_short_url":
+        bvid = utils.get_bvid_from_url(link)
+        if bvid is None:
+            await send_edit(ctx, response, "无效的Bilibili链接")
+            logger.rp(f"{link} 为无效的链接", ctx.guild)
+            return
+    else:
+        bvid = link
+
+    # 获取Bilibili视频信息
+    info_dict = await bilibili.get_info(bvid)
+
+    # 单一视频 bilibili_single
+    if info_dict["videos"] == 1 and "ugc_season" not in info_dict:
+        await add_bilibili_audio(ctx, info_dict, "bilibili_single", 0, response=response)
+
+    # 合集视频 bilibili_collection
+    elif "ugc_season" in info_dict:
+        await add_bilibili_audio(ctx, info_dict, "bilibili_single", 0, response=response)
+
+        collection_title = info_dict["ugc_season"]["title"]
+        message = f"此视频包含在合集 **{collection_title}** 中, 是否要查看此合集？\n"
+        view = CheckBilibiliCollectionView(ctx, info_dict)
+        await ctx.respond(message, view=view)
+
+    # 分P视频 bilibili_p
+    else:
+        p_info_list = []
+        for item in info_dict["pages"]:
+            p_title = item["part"]
+            p_time_str = utils.convert_duration_to_time_str(item["duration"])
+            p_info_list.append((p_title, p_time_str))
+
+        menu_list = utils.make_playlist_page(p_info_list, 10, indent=0)
+        view = EpisodeSelectView(ctx, "bilibili_p", info_dict, menu_list)
+        await send_edit(ctx, response,
+                        f"这是一个分p视频, 请选择要播放的分p:\n{menu_list[0]}\n第[1]页，共[{len(menu_list)}]页\n已输入：",
+                        view=view)
+        return
+
+
+async def add_bilibili_audio(ctx, info_dict, audio_type, num_option: int = 0,
+                             response: Union[discord.Interaction, discord.InteractionMessage, None] = None):
     # 将Interaction转换为InteractionMessage
     if isinstance(response, discord.Interaction):
         response = await response.original_response()
@@ -622,7 +644,7 @@ async def play_bili(ctx: discord.ApplicationContext, info_dict, download_type="b
     current_guild = guild_lib.get_guild(ctx)
     current_playlist = current_guild.get_playlist()
 
-    new_audio = await audio_lib_main.download_bilibili(info_dict, download_type, num_option)
+    new_audio = await audio_lib_main.download_bilibili(info_dict, audio_type, num_option)
 
     if new_audio is not None:
 
@@ -631,14 +653,11 @@ async def play_bili(ctx: discord.ApplicationContext, info_dict, download_type="b
             await play_audio(ctx, new_audio, response=response)
 
         # 如果播放列表不为空
-        elif download_type == "bili_single":
-            if isinstance(response, discord.InteractionMessage):
-                await response.edit(f"已加入播放列表：**{new_audio.get_title()} [{new_audio.get_time_str()}]**")
-            else:
-                await ctx.send(f"已加入播放列表：**{new_audio.get_title()} [{new_audio.get_time_str()}]**")
+        elif audio_type == "bilibili_single":
+            await send_edit(ctx, response, f"已加入播放列表：**{new_audio.get_title()} [{new_audio.get_time_str()}]**")
 
         current_playlist.append_audio(new_audio)
-        logger.rp(f"歌曲 {new_audio.get_title()} [{new_audio.get_time_str()}] 已加入播放列表", ctx.guild)
+        logger.rp(f"音频 {new_audio.get_title()} [{new_audio.get_time_str()}] 已加入播放列表", ctx.guild)
 
     return audio
 
@@ -663,9 +682,9 @@ async def play_ytb(ctx, url, info_dict, download_type="ytb_single",
     current_guild = guild_lib.get_guild(ctx)
     current_playlist = current_guild.get_playlist()
 
-    audio = youtube.audio_download(url, info_dict, "./downloads/", download_type)
+    new_audio = youtube.audio_download(url, info_dict, "./downloads/", download_type)
 
-    duration_str = audio.get_time_str()
+    duration_str = new_audio.get_time_str()
 
     # 如果当前播放列表为空
     if current_playlist.is_empty() and not voice_client.is_playing():
@@ -673,24 +692,174 @@ async def play_ytb(ctx, url, info_dict, download_type="ytb_single",
         voice_client.play(
             discord.PCMVolumeTransformer(
                 discord.FFmpegPCMAudio(
-                    executable=setting.value("ffmpeg_path"), source=audio._path
+                    executable=ffmpeg_path, source=new_audio.get_path()
                 )
             ), after=lambda e: asyncio.run_coroutine_threadsafe(
                 play_next(ctx), bot.loop)
         )
         # voice_client.source.volume = volume_dict[ctx.guild.id] / 100.0
 
-        logger.rp(f"开始播放：{audio._title} [{duration_str}] {audio._path}", ctx.guild)
-        await ctx.send(f"正在播放：**{audio._title} [{duration_str}]**")
+        logger.rp(f"开始播放：{new_audio.get_title()} [{duration_str}] {new_audio.get_path()}", ctx.guild)
+        await ctx.send(f"正在播放：**{new_audio.get_title()} [{duration_str}]**")
 
     # 如果播放列表不为空
     elif download_type == "ytb_single":
-        await ctx.send(f"已加入播放列表：**{audio._title} [{duration_str}]**")
+        await ctx.send(f"已加入播放列表：**{new_audio.get_title()} [{duration_str}]**")
 
-    current_playlist.append_audio(audio)
-    logger.rp(f"歌曲 {audio._title} [{duration_str}] 已加入播放列表", ctx.guild)
+    current_playlist.append_audio(new_audio)
+    logger.rp(f"歌曲 {new_audio.get_title()} [{duration_str}] 已加入播放列表", ctx.guild)
 
-    return audio
+    return new_audio
+
+
+async def search_ytb(ctx, input_name):
+    await ctx.respond("搜索功能暂未开放")
+    # name = input_name.strip()
+    #
+    # if name == "":
+    #     ctx.respond("请输入要搜索的名称")
+    #     return
+    #
+    # options = []
+    # search_result = VideosSearch(name, limit=5)
+    # info_dict = dict(search_result.result())['result']
+    #
+    # message = f"Youtube搜索 **{name}** 结果为:\n"
+    #
+    # counter = 1
+    # for result_video in info_dict:
+    #
+    #     title = result_video["title"]
+    #     video_id = result_video["id"]
+    #     duration = result_video["duration"]
+    #
+    #     options.append([title, video_id, duration])
+    #     message = message + f"**[{counter}]** {title}  [{duration}]\n"
+    #
+    #     counter += 1
+    #
+    # logger.rp(ctx, f"搜索结果为：{options}")
+    #
+    # message = message + "\n请选择："
+    #
+    # if len(info_dict) == 0:
+    #     await ctx.respond("没有搜索到任何结果")
+    #     return
+    #
+    # view = SearchSelectView(ctx, options)
+    # await ctx.respond(message, view=view)
+
+
+@bot.command(description="暂停正在播放的音频")
+async def pause(ctx):
+    """
+    暂停播放
+
+    :param ctx: 指令原句
+    :return:
+    """
+    if not await command_check(ctx):
+        return
+
+    guild_lib.check(ctx)
+
+    voice_client = ctx.guild.voice_client
+
+    if voice_client is not None and voice_client.is_playing():
+        if ctx.author.voice and voice_client.channel == ctx.author.voice.channel:
+            voice_client.pause()
+            logger.rp("暂停播放", ctx.guild)
+            await ctx.respond("暂停播放")
+        else:
+            logger.rp(f"收到pause指令时指令发出者 {ctx.author} 不在机器人所在的频道", ctx.guild)
+            await ctx.respond(f"您不在{setting.value('bot_name')}所在的频道")
+
+    else:
+        logger.rp("收到pause指令时机器人未在播放任何音乐", ctx.guild)
+        await ctx.respond("未在播放任何音乐")
+
+
+@bot.command(description="继续播放暂停的或被意外中断的音频", aliases=["restart"])
+async def resume(ctx, play_call=False):
+    """
+    恢复播放
+
+    :param ctx: 指令原句
+    :param play_call: 是否是由play指令调用来尝试恢复播放
+    :return:
+    """
+    if not await command_check(ctx):
+        return
+
+    guild_lib.check(ctx)
+
+    voice_client = ctx.guild.voice_client
+    current_guild = guild_lib.get_guild(ctx)
+    current_playlist = current_guild.get_playlist()
+
+    # 未加入语音频道的情况
+    if voice_client is None:
+        if not play_call:
+            logger.rp("收到resume指令时机器人没有加入任何语音频道", ctx.guild)
+            await ctx.respond(f"{setting.value('bot_name')}尚未加入任何语音频道")
+
+    # 被暂停播放的情况
+    elif voice_client.is_paused():
+        if ctx.author.voice and voice_client.channel == ctx.author.voice.channel:
+            voice_client.resume()
+            logger.rp("恢复播放", ctx.guild)
+            await ctx.respond("恢复播放")
+        else:
+            logger.rp(f"收到pause指令时指令发出者 {ctx.author} 不在机器人所在的频道", ctx.guild)
+            await ctx.respond(f"您不在{setting.value('bot_name')}所在的频道")
+
+    # 没有被暂停并且正在播放的情况
+    elif voice_client.is_playing():
+        if not play_call:
+            logger.rp("收到resume指令时机器人正在播放音乐", ctx.guild)
+            await ctx.respond("当前正在播放音乐")
+
+    # 没有被暂停，没有正在播放，并且播放列表中存在歌曲的情况
+    elif not current_playlist.is_empty():
+        current_audio = current_playlist.get_audio(0)
+
+        await play_audio(ctx, current_audio, function_call=True)
+
+        logger.rp("恢复中断的播放列表", ctx.guild)
+        await ctx.respond(f"恢复上次中断的播放列表")
+
+    else:
+        if not play_call:
+            logger.rp("收到resume指令时机器人没有任何被暂停的音乐", ctx.guild)
+            await ctx.respond("当前没有任何被暂停的音乐")
+
+
+@bot.command(description="显示当前播放列表")
+async def list(ctx):
+    """
+    将当前服务器播放列表发送到服务器文字频道中
+
+    :param ctx: 指令原句
+    :return:
+    """
+    if not await command_check(ctx):
+        return
+
+    guild_lib.check(ctx)
+
+    current_guild = guild_lib.get_guild(ctx)
+    current_playlist = current_guild.get_playlist()
+
+    if current_playlist.is_empty():
+        await ctx.respond("当前播放列表为空")
+    else:
+        playlist_list = utils.make_playlist_page(current_playlist.get_list_info(), 10, indent=2)
+
+        view = PlaylistMenu(ctx, current_playlist)
+        await ctx.respond(
+            content=">>> **播放列表**\n\n" + playlist_list[0] + f"\n第[1]页，共[{len(playlist_list)}]页\n",
+            view=view
+        )
 
 
 @bot.command(description=f"调整 {bot_name} 的语音频道音量")
@@ -699,7 +868,6 @@ async def volume(ctx: discord.ApplicationContext, volume_num=None) -> None:
         return
 
     guild_lib.check(ctx)
-    voice_client = ctx.guild.voice_client
     current_volume = guild_lib.get_guild(ctx).get_voice_volume()
 
     if volume_num is None:
@@ -746,3 +914,591 @@ async def volume(ctx: discord.ApplicationContext, volume_num=None) -> None:
             guild_lib.get_guild(ctx).set_voice_volume(current_volume)
             logger.rp(f"用户 {ctx.author} 已将音量设置为 {current_volume}%", ctx.guild)
             await ctx.respond(f"将音量设置为 **{current_volume}%**")
+
+
+# TODO 添加列表总时长显示
+class PlaylistMenu(View):
+
+    def __init__(self, ctx, playlist_1: playlist.Playlist, timeout: int = 60):
+        super().__init__(timeout=timeout)
+        self.ctx = ctx
+        self.playlist = playlist_1
+        self.playlist_pages = utils.make_playlist_page(self.playlist.get_list_info(), 10, indent=2)
+        self.page_num = 0
+        self.occur_time = utils.time()
+        # 保留当前第一页作为超时后显示的内容
+        self.first_page = f">>> **{self.playlist.get_name()}**\n\n{self.playlist_pages[0]}\n" \
+                          f"第[1]页，共[{len(self.playlist_pages)}]页\n"
+
+    def refresh_pages(self):
+        self.playlist_pages = utils.make_playlist_page(self.playlist.get_list_info(), 10, indent=2)
+        self.first_page = f">>> **{self.playlist.get_name()}**\n\n{self.playlist_pages[0]}\n" \
+                          f"第[1]页，共[{len(self.playlist_pages)}]页\n"
+
+    @discord.ui.button(label="上一页", style=discord.ButtonStyle.grey,
+                       custom_id="button_previous")
+    async def button_previous_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.refresh_pages()
+        # 翻页
+        if self.page_num == 0:
+            pass
+        else:
+            self.page_num -= 1
+        await msg.edit_message(
+            content=f">>> **{self.playlist.get_name()}**\n\n{self.playlist_pages[self.page_num]}\n"
+                    f"第[{self.page_num + 1}]页，"
+                    f"共[{len(self.playlist_pages)}]页\n", view=self
+        )
+
+    @discord.ui.button(label="下一页", style=discord.ButtonStyle.grey,
+                       custom_id="button_next")
+    async def button_next_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.refresh_pages()
+        # 翻页
+        if self.page_num == len(self.playlist_pages) - 1:
+            pass
+        else:
+            self.page_num += 1
+        await msg.edit_message(
+            content=f">>> **{self.playlist.get_name()}**\n\n{self.playlist_pages[self.page_num]}\n"
+                    f"第[{self.page_num + 1}]页，"
+                    f"共[{len(self.playlist_pages)}]页\n", view=self
+        )
+
+    @discord.ui.button(label="刷新", style=discord.ButtonStyle.grey,
+                       custom_id="button_refresh")
+    async def button_refresh_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.refresh_pages()
+
+        await msg.edit_message(
+            content=f">>> **{self.playlist.get_name()}**\n\n{self.playlist_pages[self.page_num]}\n"
+                    f"第[{self.page_num + 1}]页，"
+                    f"共[{len(self.playlist_pages)}]页\n", view=self
+        )
+
+    @discord.ui.button(label="关闭", style=discord.ButtonStyle.grey,
+                       custom_id="button_close")
+    async def button_close_callback(self, button, interaction):
+        button.disabled = True
+        msg = interaction.response
+        self.clear_items()
+        await msg.edit_message(content="已关闭", view=self)
+        await self.ctx.delete()
+
+    async def on_timeout(self):
+        self.clear_items()
+        await self.ctx.edit(content=self.first_page, view=self)
+        logger.rp(f"{self.occur_time}生成的播放列表菜单已超时(超时时间为{self.timeout}秒)", self.ctx.guild)
+
+
+class EpisodeSelectView(View):
+
+    def __init__(self, ctx, source, info_dict, menu_list, timeout=60):
+        """
+        初始化分集选择菜单
+
+        :param ctx: 指令原句
+        :param source: 播放源的种类（bilibili_p, bilibili_collection, youtube_playlist)
+        :param info_dict: 播放源的信息字典
+        :param menu_list: 选择菜单的文本（使用make_menu_list获取）
+        :param timeout: 超时时间（单位：秒）
+        :return:
+        """
+        super().__init__(timeout=timeout)
+        self.ctx = ctx
+        self.source = source
+        self.info_dict = info_dict
+        self.menu_list = menu_list
+        self.page_num = 0
+        self.result = []
+        self.dash_finish = True
+        self.occur_time = utils.time()
+        self.voice_client = self.ctx.guild.voice_client
+        self.current_playlist = guild_lib.get_guild(self.ctx).get_playlist()
+
+        self.finish = False
+
+    @discord.ui.button(label="1", style=discord.ButtonStyle.grey,
+                       custom_id="button_1", row=1)
+    async def button_1_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.result.append("1")
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="2", style=discord.ButtonStyle.grey,
+                       custom_id="button_2", row=1)
+    async def button_2_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.result.append("2")
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="3", style=discord.ButtonStyle.grey,
+                       custom_id="button_3", row=1)
+    async def button_3_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.result.append("3")
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="4", style=discord.ButtonStyle.grey,
+                       custom_id="button_4", row=2)
+    async def button_4_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.result.append("4")
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="5", style=discord.ButtonStyle.grey,
+                       custom_id="button_5", row=2)
+    async def button_5_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.result.append("5")
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="6", style=discord.ButtonStyle.grey,
+                       custom_id="button_6", row=2)
+    async def button_6_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.result.append("6")
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="7", style=discord.ButtonStyle.grey,
+                       custom_id="button_7", row=3)
+    async def button_7_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.result.append("7")
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="8", style=discord.ButtonStyle.grey,
+                       custom_id="button_8", row=3)
+    async def button_8_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.result.append("8")
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="9", style=discord.ButtonStyle.grey,
+                       custom_id="button_9", row=3)
+    async def button_9_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.result.append("9")
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="-", style=discord.ButtonStyle.grey,
+                       custom_id="button_dash", row=4)
+    async def button_dash_callback(self, button, interaction):
+        button.disabled = False
+
+        if len(self.result) == 0 or not \
+                self.result[len(self.result) - 1].isdigit() or not \
+                self.dash_finish:
+            return
+
+        msg = interaction.response
+        self.result.append("-")
+        self.dash_finish = False
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="0", style=discord.ButtonStyle.grey,
+                       custom_id="button_0", row=4)
+    async def button_0_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.result.append("0")
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label=",", style=discord.ButtonStyle.grey,
+                       custom_id="button_comma", row=4)
+    async def button_comma_callback(self, button, interaction):
+        button.disabled = False
+
+        if len(self.result) == 0 or not \
+                self.result[len(self.result) - 1].isdigit():
+            return
+
+        msg = interaction.response
+        self.result.append(",")
+        self.dash_finish = True
+        num = ""
+        for i in self.result:
+            num = num + i
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="<-", style=discord.ButtonStyle.grey,
+                       custom_id="button_backspace", row=1)
+    async def button_backspace_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        if len(self.result) <= 0:
+            pass
+        else:
+            self.result.pop()
+            num = ""
+            for i in self.result:
+                num = num + i
+            await msg.edit_message(
+                content=f"{self.menu_list[self.page_num]}\n第"
+                        f"[{self.page_num + 1}]页，共[{len(self.menu_list)}]页\n"
+                        f"已输入：" + num, view=self)
+
+    @discord.ui.button(label="取消", style=discord.ButtonStyle.red,
+                       custom_id="button_cancel", row=3)
+    async def button_cancel_callback(self, button, interaction):
+        button.disabled = True
+        self.finish = True
+        msg = interaction.response
+        self.clear_items()
+        await msg.edit_message(content=f"已取消", view=self)
+
+    @discord.ui.button(label="确定", style=discord.ButtonStyle.green,
+                       custom_id="button_confirm", row=4)
+    async def button_confirm_callback(self, button, interaction):
+        message = "已选择"
+        button.disabled = False
+        self.finish = True
+
+        if len(self.result) == 0 or self.result[len(self.result) - 1] == "-":
+            return
+
+        msg = interaction.response
+        self.clear_items()
+        num = ""
+        final_result = []
+        temp = []
+        for item in self.result:
+            if item == "-":
+                temp.append(int(num))
+                temp.append("-")
+                num = ""
+            elif item == ",":
+                if not len(temp) == 0 and temp[len(temp) - 1] == "-":
+                    temp.pop()
+                    start_1 = temp.pop()
+                    if int(num) < start_1:
+                        start_num = int(num)
+                        for i in range(start_1, int(num) - 1, -1):
+                            if i == 0:
+                                start_num = 1
+                            else:
+                                final_result.append(i)
+                        message = message + f"第[{start_num}]首至第[{start_1}]首(倒序)，"
+                    else:
+                        start_num = start_1
+                        for i in range(start_1, int(num) + 1):
+                            if i == 0:
+                                start_num = 1
+                            else:
+                                final_result.append(i)
+                        message = message + f"第[{start_num}]首至第[{int(num)}]首，"
+                    num = ""
+                else:
+                    if int(num) == 0:
+                        pass
+                    else:
+                        final_result.append(int(num))
+                        message = message + f"第[{int(num)}]首，"
+                    num = ""
+            else:
+                num = num + item
+        if num == "":
+            pass
+        elif not len(temp) == 0 and temp[len(temp) - 1] == "-":
+            temp.pop()
+            start_1 = temp.pop()
+            if int(num) < start_1:
+                start_num = int(num)
+                for i in range(start_1, int(num) - 1, -1):
+                    if i == 0:
+                        start_num = 1
+                    else:
+                        final_result.append(i)
+                message = message + f"第[{start_num}]首至第[{start_1}]首(倒序)，"
+            else:
+                start_num = start_1
+                for i in range(start_1, int(num) + 1):
+                    if i == 0:
+                        start_num = 1
+                    else:
+                        final_result.append(i)
+                message = message + f"第[{start_num}]首至第[{int(num)}]首，"
+        else:
+            if int(num) == 0:
+                pass
+            else:
+                final_result.append(int(num))
+                message = message + f"第[{int(num)}]首，"
+
+        if self.source == "bilibili_p":
+            for num in final_result:
+                if num > len(self.info_dict["pages"]):
+                    self.clear_items()
+                    await msg.edit_message(content="选择中含有无效分p号", view=self)
+                    return
+        elif self.source == "bilibili_collection":
+            for num in final_result:
+                if num > len(self.info_dict["ugc_season"]["sections"][0][
+                                 "episodes"]):
+                    self.clear_items()
+                    await msg.edit_message(content="选择中含有无效集数", view=self)
+                    return
+        elif self.source == "ytb_playlist":
+            for num in final_result:
+                if num > len(self.info_dict["entries"]):
+                    self.clear_items()
+                    await msg.edit_message(content="选择中含有无效集数", view=self)
+                    return
+
+        message = message[:-1]
+        self.clear_items()
+        await msg.edit_message(content=message, view=self)
+
+        await self.play_select(final_result)
+
+    @discord.ui.button(label="上一页", style=discord.ButtonStyle.blurple,
+                       custom_id="button_previous", row=3)
+    async def button_previous_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        num = ""
+        for i in self.result:
+            num = num + i
+        # 翻页
+        if self.page_num == 0:
+            return
+        else:
+            self.page_num -= 1
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="下一页", style=discord.ButtonStyle.blurple,
+                       custom_id="button_next", row=4)
+    async def button_next_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        num = ""
+        for i in self.result:
+            num = num + i
+        # 翻页
+        if self.page_num == len(self.menu_list) - 1:
+            return
+        else:
+            self.page_num += 1
+        await msg.edit_message(
+            content=f"{self.menu_list[self.page_num]}\n第[{self.page_num + 1}]页，"
+                    f"共[{len(self.menu_list)}]页\n已输入：" + num, view=self)
+
+    @discord.ui.button(label="全部", style=discord.ButtonStyle.blurple,
+                       custom_id="button_all", row=2)
+    async def button_all_callback(self, button, interaction):
+        button.disabled = False
+        self.finish = True
+        msg = interaction.response
+
+        total_num = 0
+        final_result = []
+        if self.source == "bilibili_p":
+            total_num = len(self.info_dict["pages"])
+        elif self.source == "bilibili_collection":
+            total_num = len(
+                self.info_dict["ugc_season"]["sections"][0]["episodes"])
+        elif self.source == "ytb_playlist":
+            total_num = len(self.info_dict["entries"])
+
+        for num in range(1, total_num + 1):
+            final_result.append(num)
+
+        self.clear_items()
+        await msg.edit_message(content=f"已选择全部[{total_num}]首", view=self)
+        await self.play_select(final_result)
+
+    async def play_select(self, final_result):
+        total_num = len(final_result)
+        total_duration = 0
+
+        # ----- 下载并播放视频 -----
+
+        # 如果为Bilibili分p视频
+        if self.source == "bilibili_p":
+            counter = 1
+            for item in self.info_dict["pages"]:
+                if counter in final_result:
+                    total_duration += item["duration"]
+                counter += 1
+            total_duration = utils.convert_duration_to_time_str(total_duration)
+            loading_msg = await self.ctx.send(f"正在将{total_num}首歌曲加入播放列表  总时长 -> [{total_duration}]")
+
+            for num_p in final_result:
+                await add_bilibili_audio(self.ctx, self.info_dict, "bilibili_p", num_p - 1)
+
+            await loading_msg.delete()
+            await self.ctx.send(f"已将{total_num}首歌曲加入播放列表  "
+                                f"总时长 -> [{total_duration}]")
+
+        # 如果为Bilibili合集视频
+        elif self.source == "bilibili_collection":
+            counter = 1
+            for item in self.info_dict["ugc_season"]["sections"][0]["episodes"]:
+                if counter in final_result:
+                    total_duration += item["arc"]["duration"]
+                counter += 1
+            total_duration = utils.convert_duration_to_time_str(total_duration)
+            loading_msg = await self.ctx.send(f"正在将{total_num}首歌曲加入播放"
+                                              f"列表  总时长 -> [{total_duration}]")
+
+            for num in final_result:
+                bvid = self.info_dict["ugc_season"]["sections"][0]["episodes"][
+                    num - 1]["bvid"]
+                info_dict = await bilibili.get_info(bvid)
+                await add_bilibili_audio(self.ctx, info_dict, "bilibili_collection")
+
+            await loading_msg.delete()
+            await self.ctx.send(f"已将{total_num}首歌曲加入播放列表  "
+                                f"总时长 -> [{total_duration}]")
+
+        # # 如果为Youtube播放列表
+        # elif self.source == "ytb_playlist":
+        #     counter = 1
+        #     for item in self.info_dict["entries"]:
+        #         if counter in final_result:
+        #             total_duration += item["duration"]
+        #         counter += 1
+        #     total_duration = utils.convert_duration_to_time_str(total_duration)
+        #     loading_msg = await self.ctx.send(f"正在将{total_num}首歌曲加入播放"
+        #                                       f"列表  总时长 -> [{total_duration}]")
+        #
+        #     for num in final_result:
+        #         url = f"https://www.youtube.com/watch?v=" \
+        #               f"{self.info_dict['entries'][num - 1]['id']}"
+        #         download_type, info_dict = ytb_get_info(url)
+        #         await play_ytb(self.ctx, url, info_dict, "ytb_playlist")
+        #
+        #     await loading_msg.delete()
+        #     await self.ctx.send(f"已将{total_num}首歌曲加入播放列表  "
+        #                         f"总时长 -> [{total_duration}]")
+
+        else:
+            logger.rp("未知的播放源", self.ctx.guild)
+
+    async def on_timeout(self):
+        self.clear_items()
+        if self.finish:
+            await self.ctx.edit(view=self)
+        else:
+            await self.ctx.edit(content="分集选择菜单已超时", view=self)
+        logger.rp(f"{self.occur_time}生成的搜索选择菜单已超时(超时时间为{self.timeout}秒)", self.ctx.guild)
+
+
+class CheckBilibiliCollectionView(View):
+
+    def __init__(self, ctx, info_dict, timeout=10):
+        super().__init__(timeout=timeout)
+        self.ctx = ctx
+        self.info_dict = info_dict
+        self.occur_time = utils.time()
+        self.finish = False
+
+    @discord.ui.button(label="确定", style=discord.ButtonStyle.grey,
+                       custom_id="button_confirm")
+    async def button_confirm_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.finish = True
+        message = "这是一个合集, 请选择要播放的视频:\n"
+        ep_info_list = []
+        for item in self.info_dict["ugc_season"]["sections"][0]["episodes"]:
+            ep_title = item["title"]
+            ep_time_str = utils.convert_duration_to_time_str(item["arc"]["duration"])
+            ep_info_list.append((ep_title, ep_time_str))
+
+        menu_list = utils.make_playlist_page(ep_info_list, 10, indent=0)
+        view = EpisodeSelectView(self.ctx, "bilibili_collection", self.info_dict, menu_list)
+        await self.ctx.send(f"{menu_list[0]}\n第[1]页，共["
+                            f"{len(menu_list)}]页\n已输入：",
+                            view=view)
+
+        await msg.edit_message(view=self)
+        await self.ctx.delete()
+
+    @discord.ui.button(label="关闭", style=discord.ButtonStyle.grey,
+                       custom_id="button_cancel")
+    async def button_close_callback(self, button, interaction):
+        button.disabled = False
+        msg = interaction.response
+        self.finish = True
+        self.clear_items()
+        await msg.edit_message(content="", view=self)
+
+    async def on_timeout(self):
+        if not self.finish:
+            await self.ctx.delete()
+        logger.rp(f"{self.occur_time}生成的合集查看选择栏已超时(超时时间为{self.timeout}秒)", self.ctx.guild)
